Reconfigure use of CLASSIFICATION_LEVELS
It was determined that it is important to be able
to specify which caveats should be applied to which
classification. This change tweaks the CLASSIFICATION_LEVELS
setting, and makes use of the new format to ensure
only the caveats that apply to a particular classification
are able to be selected.