Increase page refresh timer on harvest
On systems with poor network responsiveness
the low refresh time causes the harvest page
to become unusable in the event that a resource
fails to be harvested. This new timeout should
strike a balance between refreshing the page
frequently enough to update the status of "IN
PROGRESS" items without breaking things in the
event of a failure.